+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## UNRELEASED
 
+### BUG FIXES
+
+* Yorc bootstrap on 4.0.0-M7 doesn't work unless an alternative download URL is provided for Yorc ([GH-561](https://github.com/ystia/yorc/issues/561))
+
+## 4.0.0-M7 (November 29, 2019)
+
 ### BREAKING CHANGES
 
 #### Changes on the REST API
+
+##### Deployment updates
 
 Until now deployments updates (which is a premium feature) were made on the same API operation than submitting a deployment with a given identifier:
 `PUT /deployments/<deployment_id>`.
@@ -19,9 +27,15 @@ premium version is required to perform a deployment update.
 
 This is tracked on ([GH-547: Refactor Deployment updates API](https://github.com/ystia/yorc/issues/547)).
 
+#### Others
+
+* API /health changed to /server/health ([GH-551](https://github.com/ystia/yorc/issues/551))
+
 ### FEATURES
 
 * Add support for using bastion hosts to provision instances with ansible and terraform ([GH-128](https://github.com/ystia/yorc/issues/128))
+* Enrich Yorc REST API with endpoints and handlers for locations management ([GH-479](https://github.com/ystia/yorc/issues/479))
+* Loading bar while bootstrap ([GH-254](https://github.com/ystia/yorc/issues/254))
 
 ### ENHANCEMENTS